changed cshifts[] to use floats instead of ints, this reduces framerate dependence (where damage flashes faded faster at lower framerates)
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@6270 d7cf8633-e32d-0410-b094-e92efae38249

+++ client.h
@@ -366,8 +366,8 @@ typedef struct scoreboard_s
 
 typedef struct cshift_s
 {
-	int		destcolor[3];
-	int		percent;		// 0-256
+	float	destcolor[3];
+	float	percent;		// 0-256
 } cshift_t;
 
 #define	CSHIFT_CONTENTS	0
